# functions/src/modules/user/model/user-profile.ts
export type UserProfile = {
    id: string,
    name: string,
    birthDate: number
    mail: string,
    profileImageUrl: string,
    // settings
    radiusVisibility: number,
    badge: number,
    pocks: number
}